five

+++ app/Http/Controllers/Admin/Akun/PetugasController.php
@@ -29,8 +29,29 @@ class PetugasController extends Controller
 
     public function store(Request $request)
     {
-        $userPath = $request->file('photo')->store('profileadmin', 'public');
 
+        $this->validate($request, [
+            'name' => 'required',
+            'username' => 'required',
+            'photo' => 'required',
+            'email' => 'required|email',
+            'keterangan' => 'required',
+            'password' => 'required',
+        ], [
+            'name.required' => 'Nama Lengkap Wajib Diisi!',
+            'username.required' => 'Username Wajib Diisi!',
+            'photo.required' => 'Foto Wajib Diisi!',
+            'email.required' => 'Email Wajib Diisi!',
+            'keterangan.required' => 'Keterangan Wajib Diisi!',
+            'password.required' => 'Password Wajib Diisi!',
+        ]);
+
+        if ($request->hasFile('photo')) {
+            $userPath = $request->file('photo')->store('profile', 'public');
+        } else {
+            // Tangani situasi jika tidak ada file yang diunggah
+            return back()->with(['gagal' => 'Tidak ada file yang diunggah!']);
+        }
         $users = User::create([
             'name' => $request->name,
             'username' => $request->username,
@@ -46,7 +67,7 @@ class PetugasController extends Controller
             return redirect()->route('petugas.index')->with(['berhasil' => 'Data Berhasil Disimpan!']);
         }else{
             //redirect dengan pesan error
-            return redirect()->route('petugas.index')->with(['error' => 'Data Gagal Disimpan!']);
+            return redirect()->route('petugas.index')->with(['gagal' => 'Data Gagal Disimpan!']);
         }
     }
 
@@ -62,16 +83,19 @@ class PetugasController extends Controller
             'username' => 'required',
             'email' => 'required|email',
             'password' => 'nullable',
+            'keterangan' => 'required',
             'level' => 'nullable',
         ]);
 
         // Mengambil foto baru jika ada
         if ($request->hasFile('photo')) {
-            // Menghapus foto pengguna sebelumnya
-            Storage::disk('public')->delete($userPath);
+            // Menghapus foto pengguna sebelumnya jika ada
+            if (!empty($userPath)) {
+                Storage::disk('public')->delete($userPath);
+            }
 
             // Menyimpan foto baru
-            $userPath = $request->file('photo')->store('profilesiswa', 'public');
+            $userPath = $request->file('photo')->store('profile', 'public');
         }
 
         // Mengupdate data pengguna
@@ -80,6 +104,7 @@ class PetugasController extends Controller
             'username' => $request->username,
             'photo' => $userPath,
             'email' => $request->email,
+            'keterangan' => $request->keterangan,
             'level' => $request->level,
         ];
 
@@ -95,7 +120,7 @@ class PetugasController extends Controller
             return redirect()->route('petugas.index')->with(['berhasil' => 'Data Berhasil Diperbarui!']);
         } else {
             // Redirect dengan pesan error
-            return redirect()->route('petugas.index')->with(['error' => 'Data Gagal Diperbarui!']);
+            return redirect()->route('petugas.index')->with(['gagal' => 'Data Gagal Diperbarui!']);
         }
     }
 
@@ -127,7 +152,7 @@ class PetugasController extends Controller
             return redirect()->route('petugas.index')->with(['berhasil' => 'Data Berhasil Dihapus!']);
         } else {
             // Redirect dengan pesan error
-            return redirect()->route('petugas.index')->with(['error' => 'Data Gagal Dihapus!']);
+            return redirect()->route('petugas.index')->with(['gagal' => 'Data Gagal Dihapus!']);
         }
     }
 }
